+++ Code/Functions/Public/Add-PsNetHostsEntry.ps1
@@ -54,6 +54,8 @@ function Add-PsNetHostsEntry {
     )
 
     begin {
+        $function = $($MyInvocation.MyCommand.Name)
+        Write-Verbose "Running $function"
     }
     
     process {