+++ Code/Functions/Public/Remove-PsNetHostsEntry.ps1
@@ -42,6 +42,8 @@ function Remove-PsNetHostsEntry {
     )
 
     begin {
+        $function = $($MyInvocation.MyCommand.Name)
+        Write-Verbose "Running $function"
     }
     
     process {