+++ PsNetTools/0.7.6/PsNetTools.psm1
@@ -1,5 +1,5 @@
 <#
-    Generated at 11/03/2019 12:13:44 by Martin Walther
+    Generated at 10/02/2019 14:12:45 by Martin Walther
     using module ..\PsNetTools\PsNetTools.psm1
 #>
 #region namespace PsNetTools
@@ -79,6 +79,7 @@ Class PsNetDig {
 
     [PsNetDigType]static dig([String] $InputString) {
         
+        [bool]     $IsIpAddress = $false
         [DateTime] $start       = Get-Date
         [Array]    $dnsreturn   = $null
         [Array]    $collection  = $null
@@ -86,17 +87,37 @@ Class PsNetDig {
         [Array]    $ipv6address = $null
         [String]   $TargetName  = $null
 
-        $dnsreturn = [System.Net.Dns]::GetHostEntry($InputString)
-        if(-not([String]::IsNullOrEmpty($dnsreturn))){
-            $TargetName = $dnsreturn.hostname
-            $collection = $dnsreturn.AddressList
+        try {
+            $InputString = [ipaddress]$InputString
+            $IsIpAddress = $true
         }
-        
+        catch {
+            $Error.Clear()
+        }
+
+        # InputType is IPv4Address
+        if($IsIpAddress){
+            $dnsreturn = [System.Net.Dns]::GetHostByAddress($InputString)
+            if(-not([String]::IsNullOrEmpty($dnsreturn))){
+                $TargetName = $dnsreturn.hostname
+                $collection = $dnsreturn.AddressList
+            }
+        }
+
+        # InputType is Hostname
+        else{
+            $dnsreturn = [System.Net.Dns]::GetHostAddressesAsync($InputString).GetAwaiter().GetResult()
+            if(-not([String]::IsNullOrEmpty($dnsreturn))){
+                $TargetName = [System.Net.Dns]::GetHostByName($InputString).Hostname
+                $collection = $dnsreturn
+            }
+        }
+
         foreach($item in $collection){
-            if($($item.AddressFamily) -eq [System.Net.Sockets.AddressFamily]::InterNetwork){
+            if($($item.AddressFamily) -eq 'InterNetwork'){
                 $ipv4address += $item.IPAddressToString
             }
-            if($($item.AddressFamily) -eq [System.Net.Sockets.AddressFamily]::InterNetworkV6){
+            if($($item.AddressFamily) -eq 'InterNetworkV6'){
                 $ipv6address += $item.IPAddressToString
             }
         }
@@ -641,10 +662,9 @@ Class PsNetAdapterType {
     [String] $OperationalStatus
     [String] $PhysicalAddres
     [String] $IpVersion
-    [bool]   $IsAPIPAActive
+    [bool]   $IsAPIPAEnabled
     [Object] $IpV4Addresses
     [Object] $IpV6Addresses
-    [String] $DNSSuffix
 
     PsNetAdapterType (
         [bool]   $Succeeded,
@@ -652,9 +672,8 @@ Class PsNetAdapterType {
         [Object] $adapter,
         [String] $IpVersion,
         [Object] $IpV4Addresses,
-        [Object] $IpV6Addresses,
-        [String] $DNSSuffix
-        ) {
+        [Object] $IpV6Addresses
+    ) {
         $this.Succeeded            = $Succeeded
         $this.Index                = $IpV4properties.Index
         $this.Name                 = $adapter.Name
@@ -663,10 +682,9 @@ Class PsNetAdapterType {
         $this.OperationalStatus    = $adapter.OperationalStatus
         $this.PhysicalAddres       = $adapter.GetPhysicalAddress().ToString() -replace '..(?!$)', '$&:'
         $this.IpVersion            = $IpVersion
-        $this.IsAPIPAActive        = $IpV4properties.IsAutomaticPrivateAddressingActive
+        $this.IsAPIPAEnabled       = $IpV4properties.IsAutomaticPrivateAddressingActive
         $this.IpV4Addresses        = $IpV4Addresses
         $this.IpV6Addresses        = $IpV6Addresses
-        $this.DNSSuffix            = $DNSSuffix
     }
 
 }
@@ -798,7 +816,6 @@ Class PsNetAdapter {
 
                     $IpVersion      = @()
                     $properties     = $adapter.GetIPProperties()
-                    $DNSSuffix      = $properties.DNSSuffix
                     $IpV4properties = $properties.GetIPv4Properties()
                     $IpV4Addresses  = @()
                     $IpV6Addresses  = @()
@@ -816,13 +833,13 @@ Class PsNetAdapter {
                 
                     foreach ($ip in $properties.UnicastAddresses) {
                         if ($ip.Address.AddressFamily -eq [System.Net.Sockets.AddressFamily]::InterNetwork){
-                            $IpV4Addresses += $ip.Address.IPAddressToString
+                            $IpV4Addresses += $ip.Address.ToString()
                         }
-                        if ($ip.Address.AddressFamily -eq [System.Net.Sockets.AddressFamily]::InterNetworkV6){
-                            $IpV6Addresses += $ip.Address.IPAddressToString
+                        if ($ip.Address.AddressFamily -eq [System.Net.Sockets.AddressFamily]::InterNetwork6){
+                            $IpV6Addresses += $ip.Address.ToString()
                         }
                     }
-                    $resultset += [PsNetAdapterType]::New($true,$IpV4properties,$adapter,$IpVersion,$IpV4Addresses,$IpV6Addresses,$DNSSuffix)
+                    $resultset += [PsNetAdapterType]::New($true,$IpV4properties,$adapter,$IpVersion,$IpV4Addresses,$IpV6Addresses)
                 }
             }
         }
@@ -870,18 +887,18 @@ Class PsNetAdapter {
     
                     foreach ($ip in $properties.UnicastAddresses) {
                         if ($ip.Address.AddressFamily -eq [System.Net.Sockets.AddressFamily]::InterNetwork){
-                            $IpV4Addresses += $ip.Address.IPAddressToString
+                            $IpV4Addresses += $ip.Address.ToString()
                         }
-                        if ($ip.Address.AddressFamily -eq [System.Net.Sockets.AddressFamily]::InterNetworkV6){
-                            $IpV6Addresses += $ip.Address.IPAddressToString
+                        if ($ip.Address.AddressFamily -eq [System.Net.Sockets.AddressFamily]::InterNetwork6){
+                            $IpV6Addresses += $ip.Address.ToString()
                         }
                     }
 
                     foreach($gateway in $properties.GatewayAddresses){
-                        if($gateway.Address.AddressFamily -eq [System.Net.Sockets.AddressFamily]::InterNetworkV6){
+                        if($gateway.Address.AddressFamily -eq 'InterNetwork6'){
                             $GatewayIpV6Addresses += $gateway.Address.IPAddressToString
                         }
-                        if($gateway.Address.AddressFamily -eq [System.Net.Sockets.AddressFamily]::InterNetwork){
+                        if($gateway.Address.AddressFamily -eq 'InterNetwork'){
                             $GatewayIpV4Addresses += $gateway.Address.IPAddressToString
                         }
                     }
@@ -908,8 +925,8 @@ Class PsNetAdapter {
                         Index                = $IpV4properties.Index
                         Mtu                  = $IpV4properties.Mtu
                         IsForwardingEnabled  = $IpV4properties.IsForwardingEnabled
-                        IsAPIPAEnabled       = $IpV4properties.IsAutomaticPrivateAddressingEnabled
-                        IsAPIPAActive        = $IpV4properties.IsAutomaticPrivateAddressingActive
+                        IsAPIPAEnabled       = $IpV4properties.IsAutomaticPrivateAddressingActive
+                        IsAPIPAActive        = $IpV4properties.IsAutomaticPrivateAddressingEnabled
                         IsDhcpEnabled        = $IpV4properties.IsDhcpEnabled
                         UsesWins             = $IpV4properties.UsesWins
                         GatewayIpV4Addresses = $GatewayIpV4Addresses
@@ -1737,38 +1754,25 @@ Class PsNetDnsClient {
 
     [PsNetDnsClientType] static GetDnsSearchSuffix([OSType]$CurrentOS){
 
-        [String]   $function           = 'GetDnsSearchSuffix()'
-        [DateTime] $start              = Get-Date
-        [Object]   $SuffixSearchList   = $null
-        [String]   $ComputerName       = $null
+        [String]   $function = 'GetDnsSearchSuffix()'
+        [DateTime] $start    = Get-Date
         [PsNetDnsClientType]$resultset = $null
 
         # For Windows only
         if($CurrentOS -eq [OSType]::Windows){
+            
             try{
-                $ComputerName     = $env:ComputerName
                 $SuffixSearchList = (Get-DnsClientGlobalSetting).SuffixSearchList
+                $duration = $([math]::round(((New-TimeSpan $($start) $(Get-Date)).TotalMilliseconds),0))
+                $resultset = [PsNetDnsClientType]::New($true,$env:ComputerName,$SuffixSearchList,$(Get-Date -f 'yyyy-MM-dd HH:mm:ss.fff'),$duration)
             }
             catch {
                 $resultset += [PsNetError]::New("$($function)()", $_)
                 $error.Clear()
             }
+
         }
 
-        # For Linux and Mac only
-        if(($CurrentOS -eq [OSType]::Linux) -or ($CurrentOS -eq [OSType]::Mac)){
-            try{
-                $ComputerName     = hostname
-                $SuffixSearchList = (Get-Content -Path '/etc/resolv.conf' | Select-String -Pattern 'search\s\S+') -replace 'search\s'
-            }
-            catch {
-                $resultset += [PsNetError]::New("$($function)()", $_)
-                $error.Clear()
-            }
-        }
-
-        $duration = $([math]::round(((New-TimeSpan $($start) $(Get-Date)).TotalMilliseconds),0))
-        $resultset = [PsNetDnsClientType]::New($true,$ComputerName,$SuffixSearchList,$(Get-Date -f 'yyyy-MM-dd HH:mm:ss.fff'),$duration)
         return $resultset
 
     }